### 2024-03-09 #6456 @editorialbot
[PRE REVIEW]: GATree: Evolutionary decision tree classifier in Python
**Submitting author:** @lahovniktadej (<a href="http://orcid.org/0009-0005-9689-2991">Tadej Lahovnik</a>)
**Repository:** https://github.com/lahovniktadej/gatree
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.4
