### 2024-03-10 #6456 @arfon
@lahovniktadej – due to the relatively small size of this code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. We should get back to you sometime next week. If you want to fix the DOIs (noting that @editorialbot's suggestions are not always right), you can, then use the following commands (one at a time, as the first line of a new comment) to regenerate the PDF and check the references.
