+++ CONTRIBUTING.md
@@ -1,59 +1,113 @@
-# Contributing to GATree
-:+1::tada: First off, thanks for taking the time to contribute! :tada::+1:
+# Contributing
+Contributions are welcome and greatly appreciated! Please follow the guidelines below to make the process smooth and efficient for everyone involved.
 
-## Code of Conduct
-This project and everyone participating in it is governed by the [GATree Code of Conduct](CODE_OF_CONDUCT.md). By participating, you are expected to uphold this code. Please report unacceptable behavior to [tadej.lahovnik1@um.si](mailto:tadej.lahovnik1@um.si).
+## Table of Contents
+- [Contributing](#contributing)
+    - [Table of Contents](#table-of-contents)
+    - [Types of Contributions](#types-of-contributions)
+        - [Bug Reports](#bug-reports)
+        - [Feature Requests](#feature-requests)
+        - [Bug Fixes](#bug-fixes)
+        - [Feature Implementation](#feature-implementation)
+        - [Documentation](#documentation)
+    - [Pull Request Guidelines](#pull-request-guidelines)
+        - [Authoring](#authoring)
+        - [Reviewing](#reviewing)
+        - [Merging](#merging)
+    - [Local Environment for Development](#local-environment-for-development)
+        - [Documentation](#documentation-1)
+            - [Local Development](#local-development)
+            - [Build](#build)
+            - [Deployment](#deployment)
+        - [Testing](#testing)
 
-## How Can I Contribute?
+## Types of Contributions
+### Bug Reports
+If you find a bug, please open an issue and tag it with `bug`. The issue should include:
+- your operating system name and version,
+- GATree version or commit hash,
+- a short, descriptive title,
+- details steps required to reproduce the bug, and
+- details about your local environment that might be helpful in troubleshooting the issue (e.g. Python version, etc.).
 
-### Reporting Bugs
-Before creating bug reports, please check existing issues list as you might find out that you don't need to create one. When you are creating a bug report, please include as many details as possible in the [issue template](.github/templates/ISSUE_TEMPLATE.md).
+Please use [Markdown code blocks](https://help.github.com/articles/creating-and-highlighting-code-blocks) when posting Python stack traces to improve readability.
 
-### Suggesting Enhancements
+### Feature Requests
+If you'd like to request a new feature, please open an issue and tag it with `enhancement`. The issue should:
+- explain the new functionality in detail,
+- provide a use case for the new functionality, and
+- keep the scope as narrow as possible.
 
-Open new issue using the [feature request template](.github/templates/FEATURE_REQUEST.md).
+### Bug Fixes
+Look through the GitHub issues. Issues tagged with `bug` are open to whoever wants to implement them.
 
-### Pull requests
+### Feature Implementation
+Look through the GitHub issues. Issues tagged with `enhancement` are open to whoever wants to implement them.
 
-Fill in the [pull request template](.github/templates/PULL_REQUEST.md) and make sure your code is documented.
+### Documentation
+GATree could always use more documentation, whether as part of the official GATree docs, in docstrings, or even on the web in blog posts, articles, and such.
 
-## Setup development environment
+## Pull Request Guidelines
+### Authoring
+- Fill in all sections of the PR template.
+- Include a descriptive title.
+- Include a summary of the changes and the impact of the changes.
+- **Dependencies**: Be mindful about adding new dependencies and avoid unnecessary dependencies.
+    - If a new dependency is required, it should be added to `pyproject.toml`.
+- **Tests**: The pull request should include tests for the changes made. Ensure that the tests pass before submitting the pull request. See [Testing](#testing) for more information on how to run tests.
[…]
