+++ paper/paper.bib
@@ -6,6 +6,7 @@
   number={3},
   pages={291--312},
   year={2012},
+  doi={10.1109/TSMCC.2011.2157494},
   publisher={IEEE}
 }
 
@@ -16,7 +17,7 @@
   pages={101006},
   year={2022},
   issn={2210-6502},
-  doi={https://doi.org/10.1016/j.swevo.2021.101006},
+  doi={10.1016/j.swevo.2021.101006},
   url={https://www.sciencedirect.com/science/article/pii/S2210650221001681},
   author={Rafael Rivera-Lopez and Juana Canul-Reich and Efrén Mezura-Montes and Marco Antonio Cruz-Chávez},
   keywords={Machine learning, Single-solution-based metaheuristics, Evolutionary algorithms, Swarm intelligence methods},
@@ -29,6 +30,7 @@
   booktitle={Genetic Programming: European Conference, EuroGP 2000, Edinburgh, Scotland, UK, April 15-16, 2000. Proceedings 3},
   pages={247--258},
   year={2000},
+  doi={10.1007/978-3-540-46239-2_18},
   organization={Springer}
 }
 
@@ -38,6 +40,7 @@
   booktitle={International Conference on Parallel Problem Solving from Nature},
   pages={124--128},
   year={1990},
+  doi={10.1007/BFb0029742},
   organization={Springer}
 }
 
@@ -46,31 +49,15 @@
   author={Karakati{\v{c}}, Sa{\v{s}}o and Podgorelec, Vili},
   booktitle={Proceedings of the Genetic and Evolutionary Computation Conference Companion},
   pages={165--166},
-  year={2018}
+  year={2018},
+  doi={10.1145/3205651.3205774}
 }
 
 @inproceedings{baeta2021tensorgp,
-	title = {TensorGP -- Genetic Programming Engine in TensorFlow},
-	author = {Baeta, Francisco and Correia, Jo{\~{a}}o and Martins, Tiago and Machado, Penousal},
-	booktitle = {Applications of Evolutionary Computation - 24th International Conference, EvoApplications 2021},
-	year = {2021},
+	title={TensorGP -- Genetic Programming Engine in TensorFlow},
+	author={Baeta, Francisco and Correia, Jo{\~{a}}o and Martins, Tiago and Machado, Penousal},
+	booktitle={Applications of Evolutionary Computation - 24th International Conference, EvoApplications 2021},
+	year={2021},
+  doi={10.1007/978-3-030-72699-7_48},
 	organization={Springer}
-}
-
-@misc{Sipper2019tinyGP,
-  author = {Sipper, M.},
-  title = {Tiny Genetic Programming in Python},
-  year = {2019},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  howpublished = {\url{https://github.com/moshesipper/tiny_gp} }
-}
-
-@misc{gplearn,
-  author = {Stephens, Trevor},
-  title = {gplearn},
-  year = {2019},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  howpublished = {\url{https://github.com/trevorstephens/gplearn} }
 }
