### 2024-05-02 #6456 @lwu9
Hi,

I don't have enough bandwidth till May 23rd. If you still need people to
review at that time, I'm happy to help!

Best,
Lili

On Mon, 29 Apr 2024 at 12:20, Kelly L. Rowland ***@***.***>
wrote:

> hi @lwu9 <https://github.com/lwu9> @TahiriNadia
> <https://github.com/TahiriNadia> 👋 would either of you be interested in
> and available for reviewing this JOSS submission?
>
