### 2024-05-13 #6748 @editorialbot
[REVIEW]: GATree: Evolutionary decision tree classifier in Python
**Submitting author:** @lahovniktadej (<a href="http://orcid.org/0009-0005-9689-2991">Tadej Lahovnik</a>)
**Repository:** https://github.com/lahovniktadej/gatree
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.5
