### 2024-05-27 1157502
doc: Extension of method descriptions Descriptive documentation
