### 2024-05-27 fb1869e
doc: Community guidelines Contributing, reporting issues, contact information
