### 2024-05-28 0c3b9f9
doc: Visualisations Additional visualisations (average/best fitness)
