### 2024-05-28 c8861cd
doc: .plot() Example demonstrating usage of .plot()
