### 2024-05-28 fa88e17
doc: Examples and figures Minor changes to the paper (updated examples and figures)
