### 2024-05-28 46ce559
feat: Local prediction ratio Local calculation of prediction ratio (based solely on the node itself)
