+++ gatree/gatree.py
@@ -70,7 +70,7 @@ class GATree():
         return 1 - accuracy_score(root.y_true, root.y_pred) + (0.002 * root.size())
 
     @staticmethod
-    def _predict_and_evaluate(tree, X, y, fitness_function):
+    def _predict_and_evaluate(tree, X, y, fitness_function, ratio_scope):
         """
         Evaluate a tree on a training set (in parallel).
 
@@ -86,12 +86,16 @@ class GATree():
         for j in range(X.shape[0]):
             # Predict class for current instance
             tree.predict_one(X.iloc[j], y.iloc[j])
+        tree.calculate_prediction_ratio(ratio_scope=ratio_scope)
+        for j in range(X.shape[0]):
+            # Predict class for current instance
+            tree.predict_one(X.iloc[j], y.iloc[j])
         tree.fitness = fitness_function(tree)
         return tree
 
-    def fit(self, X, y, population_size=150, max_iter=2000, mutation_probability=0.1, elite_size=1, selection_tournament_size=2):
+    def fit(self, X, y, population_size=150, max_iter=2000, mutation_probability=0.1, elite_size=1, selection_tournament_size=2, ratio_scope='global'):
         """
-        Fit a tree to a training set. The population size, maximum iterations, mutation probability, elite size, and selection tournament size can be specified.
+        Fit a tree to a training set. The population size, maximum iterations, mutation probability, elite size, selection tournament, and ratio scope size can be specified.
 
         Args:
             X (pandas.DataFrame): Training data.
@@ -101,6 +105,7 @@ class GATree():
             mutation_probability (float, optional): Probability of mutation.
             elite_size (int, optional): Number of elite trees.
             selection_tournament_size (int, optional): Number of trees in tournament.
+            ratio_scope (str, optional): Scope of prediction ratio. Can be 'global' or 'local'. Global prediction ratio is calculated for the whole tree, while local prediction ratio is calculated for each node.
 
         Returns:
             Node: The fitted tree.
@@ -127,7 +132,7 @@ class GATree():
 
             # Evaluation of population
             population = Parallel(n_jobs=self.n_jobs)(delayed(GATree._predict_and_evaluate)(
-                tree, X, y, self.fitness_function) for tree in population)
+                tree, X, y, self.fitness_function, ratio_scope) for tree in population)
 
             # Sort population by fitness
             population.sort(key=lambda x: x.fitness, reverse=False)
