### 2024-05-29 c9245b4
doc: List of dependencies
