### 2024-05-29 04e56de
doc: Genetic operators Brief description of genetic operators used by GATree
