+++ README.md
@@ -1,30 +1,53 @@
-# GATree
-[![PyPI version](https://img.shields.io/pypi/v/gatree.svg)](https://img.shields.io/pypi/v/gatree.svg)
-[![GATree](https://github.com/lahovniktadej/gatree/actions/workflows/test.yml/badge.svg)](https://github.com/lahovniktadej/gatree/actions/workflows/test.yml)
-[![Documentation status](https://readthedocs.org/projects/gatree/badge/?version=latest)](https://gatree.readthedocs.io/en/latest/?badge=latest)
-![Open issues](https://isitmaintained.com/badge/open/lahovniktadej/gatree.svg)
-![Repository size](https://img.shields.io/github/repo-size/lahovniktadej/gatree)
-![License](https://img.shields.io/github/license/lahovniktadej/gatree.svg)
+<h1 align="center">
+    GATree
+</h1>
 
-* **Free software:** MIT license
-* **Documentation**: [http://gatree.readthedocs.io](http://gatree.readthedocs.io)
-* **Python**: 3.9.x, 3.10.x
-* **Dependencies**: listed in [CONTRIBUTING.md](./CONTRIBUTING.md#dependencies)
-* **Operating systems**: Windows, Ubuntu, macOS
+<p align="center">
+    <img alt="PyPI version" src="https://img.shields.io/pypi/v/gatree.svg" />
+    <img alt="PyPI - Python Version" src="https://img.shields.io/pypi/pyversions/gatree.svg">
+    <img alt="PyPI - Downloads" src="https://img.shields.io/pypi/dm/niaaml.svg" href="https://pepy.tech/project/gatree">
+    <img alt="Downloads" src="https://pepy.tech/badge/gatree">
+    <img alt="GATree" src="https://github.com/lahovniktadej/gatree/actions/workflows/test.yml/badge.svg" />
+    <img alt="Documentation status" src="https://readthedocs.org/projects/gatree/badge/?version=latest" />
+</p>
 
-## About 📋
+<p align ="center">
+    <img alt="Repository size" src="https://img.shields.io/github/repo-size/lahovniktadej/gatree" />
+    <img alt="License" src="https://img.shields.io/github/license/lahovniktadej/gatree.svg" />
+    <img alt="GitHub commit activity" src="https://img.shields.io/github/commit-activity/w/lahovniktadej/gatree.svg">
+    <img alt="Percentage of issues still open" src="http://isitmaintained.com/badge/open/lukapecnik/niaaml.svg" href="http://isitmaintained.com/project/lahovniktadej/gatree">
+    <img alt="Average time to resolve an issue" src="http://isitmaintained.com/badge/resolution/lukapecnik/niaaml.svg" href="http://isitmaintained.com/project/lahovniktadej/gatree">
+    <img alt="GitHub contributors" src="https://img.shields.io/github/contributors/lahovniktadej/gatree.svg"/>
+</p>
+
+<p align="center">
+    <a href="#-about">📋 About </a>  •
+    <a href="#-installation">📦 Installation</a> •
+    <a href="#-usage">🚀 Usage</a> •
+    <a href="#-genetic-operators-in-gatree">🧬 Genetic Operators</a> •
+    <a href="#-community-guidelines">🫂 Community Guidelines</a> •
+    <a href="#-license">📜 License</a>
+</p>
+
+## 📋 About
 GATree is a Python library designed for implementing evolutionary decision trees using a genetic algorithm approach. The library provides functionalities for selection, mutation, and crossover operations within the decision tree structure, allowing users to evolve and optimise decision trees for various classification tasks. 🌲🧬
 
[…]
