### 2024-05-29 ec81196
doc: Consistency z -> s
