### 2024-05-29 #6748 @lahovniktadej
@WeakCha Thanks for taking the time to review our paper and repository.

For the package:
1. The list of dependencies is available in `pyproject.toml` (located in the repository's root directory) and in the [documentation](https://gatree.readthedocs.io/en/latest/dev/installation.html). We have since updated the contribution guide with a list of dependencies and provided a link to the dependency list in `README.md`.
2. Thank you for the excellent suggestion. We have taken the understanding of the behaviour of these functions for granted, which is why the descriptions of how they work are rather vague. Thanks to the suggestion, we have updated `README.md` with extended descriptions of the individual functions as well as the descriptions in the library documentation.
3. `README.md` has been updated. The "Community Guidelines" section contains contact information and a link to the contribution guide.
4. `GATree` can handle datasets with larger sample sizes and different outcomes/labels. However, it is limited to classification tasks at the moment. We plan to introduce support for regression tasks and custom operators after the publication, but until then, our primary focus is on the classification task.

For the paper:
1. Thank you for the question. As mentioned in the paper, this is not the first package to introduce genetic algorithms for tree structures to Python, but this is the first package to do this with DECISION trees for classification tasks. As mentioned, existing Python libraries use genetic algorithms to induce tree PROGRAMS (sometimes referred to as genetic programming), which can but, as the literature shows, do not achieve comparative results. The difference between the decision trees and tree programs is that in decision trees, the nodes consist of rules (i.e., age>25), but in tree programs, we get formulas/programs, which are represented in the structure of the tree (i.e., age+7*wage/2-height). Still, the state-of-the-art of traditional machine learning on tabular data (not on images, text, or other unstructured data) is done with decision trees (i.e., CART, Random Forest, Gradient Boosting, AdaBoost, Histogram Boosting). It is possible to customise existing libraries to enable decision tree usage instead of tree programs, but this is neither easy nor optimised for this usage. Thus, we believe our `GATree` still presents a valuable contribution, not yet filled with any other library.
2. Of course, `GATree` can also achieve 100% test accuracy for the `iris` data. We have updated the paper and included one of the seeds which produced such a result. To ensure that the perfect result on `iris` is achieved with almost all seeds, one needs to raise the population size or the generation limit (or both). For demonstration, however, these limits were left at a lower level to enable quick library testing.
3. Fitness is the estimation of the quality of the individual decision trees, which determines whether a decision tree survives into the next generation or not. In the current implementation, this is calculated as the combination of accuracy on the test set (preferring better/higher accuracy) and the tree size (preferring smaller, more generalisable trees). The average fitness is the actual average value of all the fitness values of the entire population. The best fitness is only the one fitness value - the one from the best individual in the population.
4. Indeed, the visualisations are not part of our package. The structure of the final decision tree (best individual among all populations) can be displayed using the `GATree.plot()` method. Figure 3 was created using an external tool based on the output of this method. Additionally, we have provided a simple example of usage (`examples/plot_decision_tree.py`). We have also updated the `iris` example (`examples/iris.py`), which now contains the code for the graphs shown in Figure 2.
