### 2024-05-30 14869f1
feat: Global prediction ratio Global calculation of prediction ratio (based on all parents of the node)
