### 2024-05-31 245395d
feat: Class change based on prediction ratio .apply_prediction_ratio() changes the class to the highest prediction ratio in the list node
