### 2024-06-03 c010de2
feat: Algorithm selection New parameter for algorithm selection (standard, balanced)
