### 2024-06-03 4aa5630
doc: Short descriptions Removal of technical details from README
