### 2024-06-03 0efde36
doc: Footnotes Explanations of values shown in Figure 2
