+++ paper/paper.md
@@ -28,9 +28,11 @@ _GATree_ is a Python library that simplifies the way decision trees are construc
 
 # Overview
 
-At the heart of _GATree_'s methodology lies the integration of genetic algorithms with decision tree construction, a process inspired by natural evolution [@koza1990concept]. This evolutionary approach begins with the random generation of an initial population of decision trees, each evaluated for their fitness in solving a given supervised task on the training data. Fitness evaluation typically considers factors such as classification accuracy and tree complexity, striving for a balance that rewards both the quality of decisions and the generalisability of the decisions [@bot2000application; @barros2012survey].
+At the heart of _GATree_'s methodology lies the integration of genetic algorithms with decision tree construction, a process inspired by natural evolution [@koza1990concept]. This evolutionary approach begins with the random generation of an initial population of decision trees, each evaluated for their fitness[^1] in solving a given supervised task on the training data. Fitness evaluation typically considers factors such as classification accuracy and tree complexity, striving for a balance that rewards both the quality of decisions and the generalisability of the decisions [@bot2000application; @barros2012survey].
 
-![Overview of the evolution process.\label{fig:ga}](./images/ga.png){ width=100% }
+[^1]: Fitness is the estimation of the quality of the individual decision trees, which determines whether a decision tree survives into the next generation or not.
+
+![Overview of the evolution process.\label{fig:ga}](./images/ga.png){ width=90% }
 
 Following the principles of natural selection, trees that perform better are more likely to contribute to the next generation, either through direct selection or by producing offspring via crossover and mutation operations. Crossover involves the exchange of genetic material (i.e., tree nodes or branches) between two parent trees, while mutation introduces random changes to a tree's structure, promoting genetic diversity within the population. This iterative process, presented in \autoref{fig:ga}, of selection, crossover, and mutation continues across generations, with the algorithm converging towards more effective decision tree solutions over time.
 
@@ -40,12 +42,14 @@ The development of decision tree classifiers has long been a focal point in mach
 
 This evolutionary strategy ensures that _GATree_ can adaptively fine-tune decision trees, exploring a broader range of potential solutions and dynamically adjusting to achieve optimal performance. Such flexibility is precious in fields where classification tasks are complex, and data can exhibit varied and unpredictable patterns. Furthermore, _GATree_'s ability to customise fitness functions allows for incorporating domain-specific knowledge into the evolutionary process, enhancing the relevance and quality of the resulting decision trees.
 
-Even though there are existing Python libraries that use various meta-heuristic approaches to from machine learning tree models (i.e., _gplearn_[^1], tinyGP[^2] and _TensorGP_ [@baeta2021tensorgp]), they use symbolic regression and not decision trees. In the broader context of machine learning and data mining, _GATree_ represents a significant advancement, offering a novel solution to the limitations of existing libraries. By integrating the principles of genetic algorithms with decision tree construction, _GATree_ not only enhances the adaptability and performance of these classifiers but also provides a rich platform for further research and development in evolutionary computing and its applications in machine learning.
[…]
