### 2024-06-07 #6748 @FlyingPumba
Hi @lahovniktadej @WeakCha. Here are my comments:

For the package:
- The instructions for installation are very clear, as well as the guide for contributing. I was able to download, install and run the package.
- I appreciate the extensive documentation to methods in the source code.
- The usage example in README is clear and intuitive for anyone who has used sci-kit learn before (same API of `fit` and `predict` methods).

For the paper:
- The paper is very well written and easy to follow.
- Good overview and introductory example on the iris dataset.

Changes needed:
- As @WeakCha mentioned, I think that the paper needs to include an example of running GATree on a more real dataset, and showing that it improves performance when compared against sci-kit default implementation and other approaches to building decision trees . I.e., the following claim in the paper needs evidence to support it: "[...]. GATree represents a significant advancement, offering a novel solution to the limitations of existing libraries. By integrating the principles of genetic algorithms with decision tree construction, GATree not only enhances adaptability and performance of these classifiers [...]".
- The fact that GATree is limited to classification tasks should be added to the paper and package's README.

Minor comments:
- README states that Python 3.9 or 3.10 is needed, but I ran the package using Python 3.11 and it seems to work.
- There are many genetic/evolutionary algorithms in the literature. I believe this paper/package is implementing Standard GA. It would be good to state this somewhere.
