### 2024-06-10 f879b2e
doc: Python version 3.11, 3.12
