### 2024-06-10 ec709f0
doc: Limitation Disclaimer for limitation to classification tasks
