+++ paper/paper.md
@@ -24,17 +24,19 @@ bibliography: paper.bib
 
 # Summary
 
-_GATree_ is a Python library that simplifies the way decision trees are constructed and optimised for classification machine learning tasks. Leveraging the principles of genetic algorithms, _GATree_ allows for the dynamic evolution of decision tree structures, providing a flexible and powerful tool for machine learning practitioners. Unlike traditional decision tree algorithms that follow a deterministic path based on statistical models or information theory, _GATree_ introduces an evolutionary process where selection, mutation, and crossover operations guide the development of optimised trees. This method enhances the adaptability and performance of decision trees and opens new possibilities for addressing complex classification problems. _GATree_ stands out as a user-friendly, highly customisable solution, enabling users to tailor fitness functions and algorithm parameters to meet specific project needs, whether in academic research or practical applications.
+_GATree_ is a Python library that simplifies the way decision trees are constructed and optimised for classification machine learning tasks[^1]. Leveraging the principles of genetic algorithms, _GATree_ allows for the dynamic evolution of decision tree structures, providing a flexible and powerful tool for machine learning practitioners. Unlike traditional decision tree algorithms that follow a deterministic path based on statistical models or information theory, _GATree_ introduces an evolutionary process where selection, mutation, and crossover operations guide the development of optimised trees. This method enhances the adaptability and performance of decision trees and opens new possibilities for addressing complex classification problems. _GATree_ stands out as a user-friendly, highly customisable solution, enabling users to tailor fitness functions and algorithm parameters to meet specific project needs, whether in academic research or practical applications.
+
+[^1]: _GATree_ is currently limited to classification tasks, with support for regression tasks planned for future releases.
 
 # Overview
 
-At the heart of _GATree_'s methodology lies the integration of genetic algorithms with decision tree construction, a process inspired by natural evolution [@koza1990concept]. This evolutionary approach begins with the random generation of an initial population of decision trees, each evaluated for their fitness[^1] in solving a given supervised task on the training data. Fitness evaluation typically considers factors such as classification accuracy and tree complexity, striving for a balance that rewards both the quality of decisions and the generalisability of the decisions [@bot2000application; @barros2012survey].
+At the heart of _GATree_'s methodology lies the integration of genetic algorithms with decision tree construction, a process inspired by natural evolution [@koza1990concept]. This evolutionary approach begins with the random generation of an initial population of decision trees, each evaluated for their fitness[^2] in solving a given supervised task on the training data. Fitness evaluation typically considers factors such as classification accuracy and tree complexity, striving for a balance that rewards both the quality of decisions and the generalisability of the decisions [@bot2000application; @barros2012survey].
 
-[^1]: Fitness is the estimation of the quality of the individual decision trees, which determines whether a decision tree survives into the next generation or not.
-
-![Overview of the evolution process.\label{fig:ga}](./images/ga.png){ width=90% }
+[^2]: Fitness is the estimation of the quality of the individual decision trees, which determines whether a decision tree survives into the next generation or not.
 
 Following the principles of natural selection, trees that perform better are more likely to contribute to the next generation, either through direct selection or by producing offspring via crossover and mutation operations. Crossover involves the exchange of genetic material (i.e., tree nodes or branches) between two parent trees, while mutation introduces random changes to a tree's structure, promoting genetic diversity within the population. This iterative process, presented in \autoref{fig:ga}, of selection, crossover, and mutation continues across generations, with the algorithm converging towards more effective decision tree solutions over time.
+
[…]
