### 2024-06-10 6710f90
doc: Standard genetic algorithm Clarification on the use of standard genetic algorithms
