### 2024-06-10 #6748 @lahovniktadej
@FlyingPumba Thanks for taking the time to review our paper and repository.

Changes needed:
- Thanks for the suggestion. We are currently executing an experiment to evaluate the performance of our algorithm under different parameter settings. This experiment is being conducted on the `adult` dataset, where we assess the performance of `GATree` across 100 independent runs to account for variability and ensure the robustness of our results. For each run, we will be recording both accuracy and F1 score. Once the experiment is complete, we will update the paper and include a table summarising these results.
- We have included a note regarding `GATree`'s limitation to classification tasks in the paper and the package's `README`.

Minor comments:
- Thanks for the comment. This is an oversight on our part. We have since updated the `README` and the documentation to include Python 3.11 and 3.12.
- Indeed, this paper/package implements standard genetic algorithms. Thank you for the excellent suggestion. We have updated the `README` and the paper to state this explicitly.
