### 2024-06-27 284e879
doc: Experiment Results of the conducted experiment and comparison against other approaches
