### 2024-06-27 #6748 @lahovniktadej
@WeakCha @FlyingPumba We have updated the paper to include the results of our experiment and comparison against other approaches for building decision trees.

@FlyingPumba We appreciate your recommendations for comparison against other approaches, such as tinyGP and TensorGP. However, after careful consideration, we believe that a direct comparison between `GATree` and tinygp/TensorGP may not be appropriate for the following reasons:
- **tinygp**: This library is specifically designed for building Gaussian Process (GP) models, not for genetic algorithms or evolutionary programming. Its focus is on Gaussian Processes, which fundamentally differ from the genetic algorithm-based approach used by `GATree`. Therefore, comparing tinygp with `GATree` would not provide meaningful insights into the performance of decision tree evolution using genetic algorithms.
- **TensorGP**: While TensorGP is a general-purpose Genetic Programming engine, it is not tailored for building decision trees. Even though it could mimic classification decision trees with heavy overriding and customisation, its focus is on programs rather than trees. This makes a direct comparison with `GATree`, specialised for evolutionary decision tree construction, less relevant and potentially misleading.
