### 2024-07-03 #6748 @FlyingPumba
Hi @lahovniktadej, thanks for the update. I agree with @WeakCha on points 1 to 3. As for point 4, I think it's OK not to compare against `tinygp` and `TensorGP` if these libraries do not build decision trees. The new experiment comparing against `scikit-learn` and `gplearn` is already a great improvement. One question that I have for that new part of the paper: I see that the `DecisionTreeClassifier` of `scikit-learn` is limited to max depth of 5, are you also limiting the depth of the decision trees built by `GATree` or `gplearn`? If not, this could be a threat to validity to the results shown.
