+++ paper/paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 _GATree_ is a Python library that simplifies the way decision trees are constructed and optimised for classification machine learning tasks[^1]. Leveraging the principles of standard genetic algorithms, _GATree_ allows for the dynamic evolution of decision tree structures, providing a flexible and powerful tool for machine learning practitioners. Unlike traditional decision tree algorithms that follow a deterministic path based on statistical models or information theory, _GATree_ introduces an evolutionary process where selection, mutation, and crossover operations guide the development of optimised trees. This method enhances the adaptability and performance of decision trees and opens new possibilities for addressing complex classification problems. _GATree_ stands out as a user-friendly, highly customisable solution, enabling users to tailor fitness functions and algorithm parameters to meet specific project needs, whether in academic research or practical applications.
 
-[^1]: _GATree_ is currently limited to classification tasks, with support for regression tasks planned for future releases.
+[^1]: _GATree_ is limited to classification tasks, with support for regression tasks planned for future releases.
 
 # Overview
 
@@ -34,9 +34,11 @@ At the heart of _GATree_'s methodology lies the integration of genetic algorithm
 
 [^2]: Fitness is the estimation of the quality of the individual decision trees, which determines whether a decision tree survives into the next generation or not.
 
-Following the principles of natural selection, trees that perform better are more likely to contribute to the next generation, either through direct selection or by producing offspring via crossover and mutation operations. Crossover involves the exchange of genetic material (i.e., tree nodes or branches) between two parent trees, while mutation introduces random changes to a tree's structure, promoting genetic diversity within the population. This iterative process, presented in \autoref{fig:ga}, of selection, crossover, and mutation continues across generations, with the algorithm converging towards more effective decision tree solutions over time.
-
 ![Overview of the evolution process.\label{fig:ga}](./images/ga.png){ width=85% }
+
+\newpage
+
+Following the principles of natural selection, trees that perform better are more likely to contribute to the next generation, either through direct selection or by producing offspring via crossover and mutation operations. Crossover involves the exchange of genetic material (i.e., tree nodes or branches) between two parent trees, while mutation introduces random changes to a tree's structure, promoting genetic diversity within the population. This iterative process of selection, crossover, and mutation (presented in \autoref{fig:ga}) continues across generations, with the algorithm converging towards more effective decision tree solutions over time.
 
 # Statement of need
 
@@ -49,9 +51,17 @@ Even though there are existing Python libraries that use various meta-heuristic 
 [^3]: \url{https://github.com/trevorstephens/gplearn}
 [^4]: \url{https://github.com/moshesipper/tiny_gp}
 
-_GATree_, a Python library with a modular and extensible architecture, comprised of two classes: _GATree_ and _Node_. The _GATree_ class is responsible for the genetic algorithm by utilising operator classes, such as _Selection_ (with optional elitism), _Crossover_, and _Mutation_. The _Node_ class handles the decision tree structure and its operations. The library is user-friendly and highly customisable - users can easily define custom fitness functions[^5] and other parameters to meet their needs. It is implemented to be compatible with the de-facto standard _scikit-learn_ machine learning library; thus, the main methods of use (i.e., _fit()_ and _predict()_) are present in _GATree_. The following example shows how to perform classification of the _iris_ dataset using the _GATree_ library.
+## Architecture
+
[…]
