### 2024-07-22 #6748 @lahovniktadej
@WeakCha Thanks for the follow-up. For the paper:
1. We have included a brief description of the `adult` dataset in our paper. The description covers sample size, features, and the outcome type.
2. The comparison methods had a much smaller F1 score because the depth of the built decision trees was limited, whereas the depth of the decision trees constructed by `GATree` was not limited. We have since reran the experiment, limited the maximum depth of all classifiers, and updated the results in the paper. Additionally, we have included the code used to conduct the experiment in `examples/joss_experiment.py`, which allows the reproduction of the results.
3. Thanks for the suggestion. We have added subsections to ensure better paper organisation (architecture, usage and customisation, and experiment). This should make the paper easier to navigate and understand.
4. After further review, we agree that as long as `tinygp` and `TensorGP` support decision tree training and prediction, they can be considered competing methods. However, these libraries do not natively provide interfaces for decision tree prediction, and significant customisation would be required to achieve similar functionality. Therefore, we have not included them in our comparison. If they include such support in the future, it would be interesting to compare our algorithm against them.

@FlyingPumba Thanks for the follow-up. This is an oversight on our part. To ensure the results are valid, we reran the experiment and set the maximum depth of all classifiers to 5. We have updated the paper and included the new results in the table.
