### 2024-07-29 0555a2b
feat: Inherit Prediction ratio calculation based on parent nodes, parameter naming
