+++ gatree/tree/node.py
@@ -68,7 +68,7 @@ class Node:
 
     def set_left(self, n):
         """
-        Sets the left child of this node to the given node. Also sets the parent of the given node to this node.
+        Sets the left child of the current node to the given node. Also sets the parent of the given node to the current node.
 
         Args:
             n (Node): Node to be set as the left child.
@@ -78,7 +78,7 @@ class Node:
 
     def set_right(self, n):
         """
-        Sets the right child of this node to the given node. Also sets the parent of the given node to this node.
+        Sets the right child of the current node to the given node. Also sets the parent of the given node to the current node.
 
         Args:
             n (Node): Node to be set as the right child.
@@ -99,7 +99,7 @@ class Node:
 
     def get_children(self):
         """
-        Returns the children of this node.
+        Returns the children of the current node.
 
         Returns:
             tuple: Tuple containing the left and right children nodes.
@@ -108,10 +108,10 @@ class Node:
 
     def get_leaves(self):
         """
-        Returns the leaves of this node.
+        Returns the leaves of the current node.
 
         Returns:
-            list: List of leaves of this node.
+            list: List of leaves of the current node.
         """
         if self.left is None:
             return [self]
@@ -119,10 +119,10 @@ class Node:
 
     def depth(self):
         """
-        Returns the depth of this node.
+        Returns the depth of the current node.
 
         Returns:
-            int: Depth of this node.
+            int: Depth of the current node.
         """
         if self.parent is None:
             return 1
@@ -225,24 +225,28 @@ class Node:
 
         return node
 
-    def clear_evaluation(self):
+    def clear_evaluation(self, ratio=False):
         """
-        Clears the evaluation of this node and all its children. The true class values and predicted class values are reset to empty lists.
+        Clears the evaluation of the current node and all its children. The true class values and predicted class values are reset to empty lists.
+
+        Args:
+            ratio (bool, optional): If true, the prediction ratio is also reset
         """
         self.y_true = []
         self.y_pred = []
-        self.prediction_ratio = {}
+        if ratio:
+            self.prediction_ratio = {}
         if self.left:
-            self.left.clear_evaluation()
+            self.left.clear_evaluation(ratio=ratio)
         if self.right:
-            self.right.clear_evaluation()
+            self.right.clear_evaluation(ratio=ratio)
 
     def is_evaluated(self):
         """
-        Returns true if this node and all its children are evaluated. A node is evaluated if it has true class values and predicted class values.
+        Returns true if the current node and all its children are evaluated. A node is evaluated if it has true class values and predicted class values.
 
         Returns:
-            bool: True if this node and all its children are evaluated.
+            bool: True if the current node and all its children are evaluated.
         """
         if len(self.y_pred) == 0 and len(self.y_true) == 0:
             return False
@@ -283,12 +287,12 @@ class Node:
             print(e)
             return -1
 
-    def calculate_prediction_ratio(self, ratio_scope, classes):
+    def calculate_prediction_ratio(self, scope, classes):
         """
-        Calculates the prediction ratio for this node. The prediction ratio is the ratio of the predicted class values to the total number of instances in the node. The prediction ratio is stored in the prediction_ratio attribute.
+        Calculates the prediction ratio for the current node. The prediction ratio is the ratio of the predicted class values to the total number of instances in the node. The prediction ratio is stored in the prediction_ratio attribute.
 
         Args:
[…]
