### 2024-08-09 #6748 @FlyingPumba
@lahovniktadej Thanks for the changes! The article is looking much better now. I don't have any more comments on it. I think it is ready for acceptance once you make the minor changes requested by @WeakCha
