### 2024-08-12 #6748 @lahovniktadej
@WeakCha @FlyingPumba Thanks for the feedback. In the comparison table, we have included the 95% confidence intervals (CI) for the average results across all metrics and parameter sets/models. We appreciate your time and effort in reviewing our paper and believe these changes enhance its quality.
