+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "gatree"
-version = "0.1.4"
+version = "0.1.5"
 description = ""
 authors = ["Tadej Lahovnik <tadej.lahovnik@student.um.si>"]
 readme = "README.md"
