### 2024-08-12 #6748 @lahovniktadej
@kellyrowland Thank you for providing the instructions. We have issued a new tagged release of the software, version [0.1.5](https://github.com/lahovniktadej/gatree/releases/tag/0.1.5), and archived it on Zenodo. The DOI for the archived version is [10.5281/zenodo.13306888](https://doi.org/10.5281/zenodo.13306888). Please let us know if there are any further steps we should follow.
