### 2024-08-12 #6748 @lahovniktadej
@kellyrowland Thank you for bringing this to our attention. Since GitHub and Zenodo integration doesn't allow metadata edits after publication, we issued a new release to update the license information. The new DOI is [10.5281/zenodo.13307403](https://zenodo.org/doi/10.5281/zenodo.13307403), and it now correctly reflects the MIT license in its metadata. Please let us know if any further adjustments are needed.
