### 2024-08-12 #6748 @crvernon
Very clean submission @lahovniktadej and all.  This one made my job easy!  I haven't often been able to move straight through to acceptance without requesting some modification.  Thanks!
