### 2024-08-19 #6748 @lahovniktadej
@crvernon Upon reviewing the published article, we realised that the acknowledgements were unintentionally omitted, and the affiliation needs to be updated to reflect the full title of our faculty. We have made the necessary revisions to the `paper.md` file. Is it possible to reaccept the paper to ensure these updates are incorporated into the final publication? We greatly appreciate your help with this.
